requirements.txt
# Core
streamlit==1.35.0
streamlit-lottie==0.0.5

# Speech & Audio
openai-whisper==20231117
pydub==0.25.1
sounddevice==0.4.6
soundfile==0.12.1
pyttsx3==2.90
gTTS==2.5.0
edge-tts>=7.0.0  # Fast Microsoft TTS (faster than gTTS)

# API & LLM
groq==0.9.0
python-dotenv==1.0.0

# Utilities
requests==2.31.0

# Optional but recommended
numpy==1.24.3
scipy==1.11.2
